% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cobra.r
\name{cobra_internal}
\alias{cobra_internal}
\title{Convex Biclustering via Dykstra-like Proximal Algorithm}
\usage{
cobra_internal(X, Lambda_row, Lambda_col, E_row, E_col, w_row, w_col, gamma,
  max_iter = 100, tol = 0.001)
}
\arguments{
\item{X}{Data matrix}

\item{Lambda_row}{Initial guess of row Langrage multipliers}

\item{Lambda_col}{Initial guess of column Langrage multipliers}

\item{E_row}{Edge-incidence matrix for row graph}

\item{E_col}{Edge-incidence matrix for column graph}

\item{w_row}{Vector of weights for row graph}

\item{w_col}{Vector of weights for column graph}

\item{gamma}{Regularization parameter for shrinkage}

\item{max_iter}{Maximum number of AMA iterations}

\item{tol}{stopping tolerance}
}
\description{
\code{cobra_internal} is an R wrapper around C code.
}

