% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.r
\name{convert}
\alias{convert}
\title{Convert weights, as COO matrix, to CSC matrix and weights vector}
\usage{
convert(W)
}
\arguments{
\item{W}{COO matrix of weights: (i,j,w[ij])}
}
\description{
\code{convert} Takes a weights triple (COO matrix) and converts it to a sparse edge-incidence matrix (CSC matrix) and weights vector.
}
\examples{
W <- matrix(0,3,3)
W[1,] <- c(1,2,1)
W[2,] <- c(1,3,2)
W[3,] <- c(2,3,3)

sol <- convert(W)
}

