% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.r
\name{create_adjacency}
\alias{create_adjacency}
\title{Create adjacency matrix from V}
\usage{
create_adjacency(V, Phi)
}
\arguments{
\item{V}{Matrix of centroid differences}

\item{Phi}{Edge-incidence matrix}
}
\description{
\code{create_adjacency} creates an n-by-n sparse adjacency matrix from the matrix of centroid differences.
}

