% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.r
\name{create_edge_incidence}
\alias{create_edge_incidence}
\title{Edge-Incidence Matrix of Weights Graph}
\usage{
create_edge_incidence(w, n)
}
\arguments{
\item{w}{Weights vector}

\item{n}{Number of points being clustered}
}
\description{
Construct the edge-incidence matrix of the weights graph.
}

