% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cobra_validate.r
\name{get_subgroup_means}
\alias{get_subgroup_means}
\title{Get Subgroup Means}
\usage{
get_subgroup_means(X, Theta, clusters_row, clusters_col)
}
\arguments{
\item{X}{Data matrix}

\item{Theta}{validation index set - vector}

\item{clusters_row}{Row cluster object}

\item{clusters_col}{Column cluster object}
}
\description{
\code{get_subgroup_means} computes the subgroup means on the training set for validation.
}
\seealso{
\code{\link[cvxbiclustr]{get_subgroup_means_full}}
}

