% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.r
\name{weights_graph}
\alias{weights_graph}
\title{Weights Graph Adjacency Matrix}
\usage{
weights_graph(w, n)
}
\arguments{
\item{w}{Weights vector}

\item{n}{Number of points being clustered}
}
\description{
Constructs the adjacency matrix of the weights graph. This is useful to determine the connectivity of the weights graph.
}

